requirements.txt
PyQt6
requests
python-xlib
